+++ minimint-api/Cargo.toml
@@ -15,6 +15,7 @@ hex = "0.4.3"
 lightning-invoice = "0.17.0"
 minimint-derive = { path = "../minimint-derive" }
 rand = "0.6.0"
+rand07 = { version = "0.7", package = "rand" }
 secp256k1-zkp = { version = "0.6.0", features = [ "use-serde", "bitcoin_hashes", "global-context" ] }
 serde = { version = "1.0.118", features = [ "derive" ] }
 serde_json = "1.0.79"
