+++ minimint/Cargo.toml
@@ -17,7 +17,7 @@ bitcoin = "0.28.1"
 bytes = "1.1.0"
 clap = { version = "3.1.18", features = ["derive"] }
 futures = "0.3.9"
-hbbft = { git = "https://github.com/fedimint/hbbft", branch = "minimint" }
+hbbft = { git = "https://github.com/fedimint/hbbft" }
 hex = "0.4.2"
 itertools = "0.10.0"
 jsonrpsee = { version = "0.14.0", features = ["ws-server"] }
