+++ modules/minimint-ln/src/config.rs
@@ -1,6 +1,7 @@
 use minimint_api::config::GenerateConfig;
+use minimint_api::rand::Rand07Compat;
 use minimint_api::PeerId;
-use secp256k1::rand::{CryptoRng, RngCore as RngCore06};
+use secp256k1::rand::{CryptoRng, RngCore};
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
@@ -18,26 +19,6 @@ pub struct LightningModuleClientConfig {
     pub threshold_pub_key: threshold_crypto::PublicKey,
 }
 
-struct Rand07Compat<R: RngCore06>(R);
-
-impl<R: RngCore06> rand07::RngCore for Rand07Compat<R> {
-    fn next_u32(&mut self) -> u32 {
-        self.0.next_u32()
-    }
-
-    fn next_u64(&mut self) -> u64 {
-        self.0.next_u64()
-    }
-
-    fn fill_bytes(&mut self, dest: &mut [u8]) {
-        self.0.fill_bytes(dest)
-    }
-
-    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), rand07::Error> {
-        self.0.try_fill_bytes(dest).map_err(rand07::Error::new)
-    }
-}
-
 impl GenerateConfig for LightningModuleConfig {
     type Params = ();
     type ClientConfig = LightningModuleClientConfig;
@@ -46,7 +27,7 @@ impl GenerateConfig for LightningModuleConfig {
         peers: &[PeerId],
         max_evil: usize,
         _params: &Self::Params,
-        rng: impl RngCore06 + CryptoRng,
+        rng: impl RngCore + CryptoRng,
     ) -> (BTreeMap<PeerId, Self>, Self::ClientConfig) {
         let threshold = peers.len() - max_evil;
         let sks = threshold_crypto::SecretKeySet::random(threshold - 1, &mut Rand07Compat(rng));
